Add script generation

+++ src/test/java/com/seleniumtests/ut/util/squashta/TestTaScriptGenerator.java
@@ -14,7 +14,6 @@ import org.testng.annotations.Test;
 
 import com.seleniumtests.core.SeleniumTestsContext;
 import com.seleniumtests.core.SeleniumTestsContextManager;
-import com.seleniumtests.util.FileUtility;
 import com.seleniumtests.util.squashta.SquashTaTestDef;
 import com.seleniumtests.util.squashta.TaScriptGenerator;
 
@@ -32,7 +31,7 @@ public class TestTaScriptGenerator {
 	@Test(groups={"squash"})
 	public void testParseFeatures() {
 		
-		TaScriptGenerator scriptGenerator = new TaScriptGenerator(SeleniumTestsContext.getRootPath(), "core");
+		TaScriptGenerator scriptGenerator = new TaScriptGenerator("core", SeleniumTestsContext.getRootPath(), SeleniumTestsContext.getRootPath());
 		List<String> scenarios = scriptGenerator.parseFeatures();
 		Assert.assertTrue(scenarios.contains("core_ <text>"));
 		Assert.assertTrue(scenarios.contains("core_3"));
@@ -45,7 +44,7 @@ public class TestTaScriptGenerator {
 	
 	@Test(groups={"squash"})
 	public void testParseTestNg() {
-		TaScriptGenerator scriptGenerator = new TaScriptGenerator(SeleniumTestsContext.getRootPath(), "core");
+		TaScriptGenerator scriptGenerator = new TaScriptGenerator("core", SeleniumTestsContext.getRootPath(), SeleniumTestsContext.getRootPath());
 		List<SquashTaTestDef> testList = scriptGenerator.parseTestNgXml();
 		Assert.assertTrue(testList.contains(new SquashTaTestDef(null, "SquashTa_cucumber_test", true, "Conge")));
 		Assert.assertTrue(testList.contains(new SquashTaTestDef(null, "SquashTa_cucumber_tags", true, "@new")));
@@ -66,7 +65,7 @@ public class TestTaScriptGenerator {
 		FileUtils.write(Paths.get(tmpFolder.getPath(), "a_file.ta").toFile(), "");
 		FileUtils.write(Paths.get(tmpFolder.getPath(), "g__a_file.ta").toFile(), "");
 		FileUtils.write(Paths.get(tmpFolder.getPath(), "g__an_other_file.ta").toFile(), "");
-		TaScriptGenerator scriptGenerator = new TaScriptGenerator(SeleniumTestsContext.getRootPath(), "core");
+		TaScriptGenerator scriptGenerator = new TaScriptGenerator("core", SeleniumTestsContext.getRootPath(), SeleniumTestsContext.getRootPath());
 		
 		try {
 			scriptGenerator.cleanGeneratedFile(tmpFolder.toString(), Arrays.asList("g__a_file.ta", "g__a_file2.ta"));
@@ -85,23 +84,58 @@ public class TestTaScriptGenerator {
 	 */
 	@Test(groups={"squash"})
 	public void testGenerateTaScripts() throws IOException {
-		TaScriptGenerator scriptGenerator = new TaScriptGenerator(SeleniumTestsContext.getRootPath(), "core");
-		File generationFolder = Paths.get(SeleniumTestsContext.getRootPath(), "data", "core", "squash-ta", "src", "squashTA", "tests").toFile();
+		File generationFolder = Paths.get(SeleniumTestsContext.getRootPath(), "data", "core", "squash-ta").toFile();
+		TaScriptGenerator scriptGenerator = new TaScriptGenerator("core", SeleniumTestsContext.getRootPath(), generationFolder.getPath());
+		
 		try {
 			scriptGenerator.generateTaScripts();
-			Assert.assertTrue(Paths.get(generationFolder.getPath(), "g__squashta.xml_SquashTa_cucumber_tags.ta").toFile().isFile());
-			Assert.assertTrue(Paths.get(generationFolder.getPath(), "g__squashta.xml_SquashTa_cucumber_test.ta").toFile().isFile());
-			Assert.assertTrue(Paths.get(generationFolder.getPath(), "g__squashta.xml_SquashTa_testPlan.ta").toFile().isFile());
-			Assert.assertTrue(Paths.get(generationFolder.getPath(), "g__tu.xml_Unit tests with context.ta").toFile().isFile());
-			Assert.assertTrue(Paths.get(generationFolder.getPath(), "g__squashta.xml_SquashTa_cucumber_generic_core__text.ta").toFile().isFile());
-			Assert.assertTrue(Paths.get(generationFolder.getPath(), "core_generic.ta").toFile().isFile());
+			Assert.assertTrue(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__squashta.xml_SquashTa_cucumber_tags.ta").toFile().isFile());
+			Assert.assertTrue(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__squashta.xml_SquashTa_cucumber_test.ta").toFile().isFile());
+			Assert.assertTrue(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__squashta.xml_SquashTa_testPlan.ta").toFile().isFile());
+			Assert.assertTrue(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__tu.xml_Unit tests with context.ta").toFile().isFile());
+			Assert.assertTrue(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__squashta.xml_SquashTa_cucumber_generic_core__text.ta").toFile().isFile());
+			Assert.assertTrue(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "core_generic.ta").toFile().isFile());
 		} finally {
-			for (File f: generationFolder.listFiles()) {
-				if (!f.getName().equals("core_generic.ta")) {
+			for (File f: Paths.get(generationFolder.getPath(), "src", "squashTA", "tests").toFile().listFiles()) {
+				if (!f.getName().equals("core_generic.ta") && f.isFile()) {
 					f.delete();
 				}
 			}
 
+		}
+	}
+	
+	/**
+	 * Test the right scripts are generated
+	 * @throws IOException 
+	 */
+	@Test(groups={"squash"})
+	public void testGenerateTaScriptContent() throws IOException {
+		File generationFolder = Paths.get(SeleniumTestsContext.getRootPath(), "data", "core", "squash-ta").toFile();
+		TaScriptGenerator scriptGenerator = new TaScriptGenerator("core", SeleniumTestsContext.getRootPath(), generationFolder.getPath());
+		
+		try {
+			// compare non generic cucumber
+			scriptGenerator.generateTaScripts();
+			String content = FileUtils.readFileToString(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__squashta.xml_SquashTa_cucumber_tags.ta").toFile());
+			Assert.assertTrue(content.contains("application:core"));
+			Assert.assertTrue(content.contains("testngFile:%STF_HOME%"));
+			Assert.assertTrue(content.contains("TC_CUF_cucumberTest:"));
+			Assert.assertTrue(content.contains("testngName:SquashTa_cucumber_tags"));
+			
+			// compare generic cucumber
+			content = FileUtils.readFileToString(Paths.get(generationFolder.getPath(), "src", "squashTA", "tests", "g__squashta.xml_SquashTa_cucumber_generic_core__text.ta").toFile());
+			Assert.assertTrue(content.contains("application:core"));
+			Assert.assertTrue(content.contains("testngFile:%STF_HOME%/data/core/testng/squashta.xml"));
+			Assert.assertTrue(content.contains("TC_CUF_cucumberTest:-DcucumberTests\"=core_&nbsp;<text>\""));
+			Assert.assertTrue(content.contains("testngName:SquashTa_cucumber_generic"));
+		} finally {
+			for (File f: Paths.get(generationFolder.getPath(), "src", "squashTA", "tests").toFile().listFiles()) {
+				if (!f.getName().equals("core_generic.ta") && f.isFile()) {
+					f.delete();
+				}
+			}
+			
 		}
 	}
 }
